Refactor and cleanup the muxing/proxy client
Also removes the config to enable/disable muxing (as the need for that
option doesn't really apply anymore)
